backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.2.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2